This is a common source of confusion, but it matches the NumPy
behavior.

Fixes #44010
Fixes #59526

Differential Revision: [D30307646](https://our.internmc.facebook.com/intern/diff/D30307646)

[ghstack-poisoned]

+++ caffe2/opt/shape_info.cc
@@ -216,8 +216,7 @@ void parseShapeInfoMapFromString(
     }
 
     bool valid = true;
-    // NOLINTNEXTLINE(clang-diagnostic-sign-compare)
-    for (int i = 1; i < size; i++) {
+    for (size_t i = 1; i < size; i++) {
       auto dim = kv[i];
       try {
         shape.add_dims(std::stoi(dim));
